Revert "evas gl: move to floating point coordinate system."
This reverts commit 2ac93b5005e0693c5d84df71b4aa1c5a4b6baa80.

+++ src/modules/evas/engines/gl_common/evas_gl_common.h
@@ -335,7 +335,7 @@ struct _Evas_Engine_GL_Context
       } shader;
       struct {
          int            num, alloc;
-         GLfloat       *vertex;
+         GLshort       *vertex;
          GLubyte       *color;
          GLfloat       *texuv;
          GLfloat       *texuv2;
